backend/data_ingress/requirements.txt
requests
beautifulsoup4
html2text
readability-lxml
undetected-chromedriver
selenium
PyPDF2